omit curly braces on single statement block

+++ controllers/threadController.js
@@ -27,9 +27,9 @@ threadController.create = async (req, res, next) => {
       user: req.user._id,
       image: !req.file ? null : `${Date.now()}-${req.file.originalname}`
     })
-    if (thread.image) {
+    if (thread.image)
       fs.writeFileSync(`./public/images/${thread.image}`, req.file.buffer)
-    }
+
     res.redirect('/threadlist')
   } catch (err) {
     next(err)
@@ -39,15 +39,11 @@ threadController.create = async (req, res, next) => {
 threadController.get = async (req, res, next) => {
   try {
     const { id } = req.params
-    if (!ObjectId.isValid(id)) {
-      return next()
-    }
+    if (!ObjectId.isValid(id)) return next()
     const thread = await Thread.findById(id)
       .populate('user')
       .populate('comments.user')
-    if (!thread) {
-      return next()
-    }
+    if (!thread) return next()
     res.render('viewThread', { thread, user: req.user, moment })
   } catch (err) {
     next(err)
@@ -57,18 +53,12 @@ threadController.get = async (req, res, next) => {
 threadController.delete = async (req, res, next) => {
   try {
     const { id } = req.params
-    if (!ObjectId.isValid(id)) {
-      return next()
-    }
+    if (!ObjectId.isValid(id)) return next()
     const thread = await Thread.findById(id)
-    if (!thread) {
-      return next()
-    }
+    if (!thread) return next()
     if (thread.user.equals(req.user._id)) {
       await Thread.findByIdAndDelete(id)
-      if (thread.image) {
-        fs.unlinkSync(`./public/images/${thread.image}`)
-      }
+      if (thread.image) fs.unlinkSync(`./public/images/${thread.image}`)
     }
     res.redirect('/threadlist')
   } catch (err) {
@@ -79,13 +69,9 @@ threadController.delete = async (req, res, next) => {
 threadController.renderEditPage = async (req, res, next) => {
   try {
     const { id } = req.params
-    if (!ObjectId.isValid(id)) {
-      return next()
-    }
+    if (!ObjectId.isValid(id)) return next()
     const thread = await Thread.findById(id).populate('user')
-    if (!thread) {
-      return next()
-    }
+    if (!thread) return next()
     res.render('editThread', { thread, errors: req.flash('errors') })
   } catch (err) {
     next(err)
@@ -95,19 +81,13 @@ threadController.renderEditPage = async (req, res, next) => {
 threadController.edit = async (req, res, next) => {
   try {
     const { id } = req.params
-    if (!ObjectId.isValid(id)) {
-      return next()
-    }
+    if (!ObjectId.isValid(id)) return next()
     let thread = await Thread.findById(id).populate('user')
-    if (!thread) {
-      return next()
-    }
+    if (!thread) return next()
     if (thread.user.equals(req.user._id)) {
       const { title } = req.body
       if (req.file) {
-        if (thread.image) {
-          fs.unlinkSync(`./public/images/${thread.image}`) // delete current image
-        }
+        if (thread.image) fs.unlinkSync(`./public/images/${thread.image}`) // delete current image
         thread = await Thread.findByIdAndUpdate(
           id,
           {
@@ -117,9 +97,7 @@ threadController.edit = async (req, res, next) => {
           { new: true }
         )
         fs.writeFileSync(`./public/images/${thread.image}`, req.file.buffer)
-      } else {
-        await Thread.findByIdAndUpdate(id, { title })
-      }
+      } else await Thread.findByIdAndUpdate(id, { title })
     }
     res.redirect(`/view/${id}`)
   } catch (err) {
